Added ODB support.
The ODB object relational mappings are now included into the project. 
I am happy to say that the code generator is integrated with the CMake building environment, so that the build process is still fully automated. A small script was included to facilitate this (in /cmake/Modules).

Also added db/PlayerAccount.hpp as a first attempt to generate bindings.

Small update to README to keep track of dependencies and how we install them.