DB fully integrated; embedded automatic schema generation works.
Also did some refactoring with typedefs.hpp - what a horible idea, FIXME
I had to do this because typedefs was giving the ODB compiler too many dependencies. 
This may become a problem, as it does not parse C++14. If need be, we can #ifndef __ODB_COMPILER__ (I think, or something similar) to fix issues.

Specifying the include path for the ODB compiler is still broken and currently only hacked in the CMakeLists.txt.
There is a huge problem with the order of CMake executing the code generators and their dependencies - if ODB needs to include *.pb.h headers we are in trouble.
However, it should be possible to keep the dependencies out of the headers. Who would want to store PB Messages in the database anyway?