Server/Client communication now works both ways.
Had to fix a problem with interpreting websock messages serialized from google protocol buffers in javascript: It kept throwing 'illegal utf-8 character' exception. This seems to be a problem whenever the protobuff library encodes something in a tricky way (like a message consisting of a bool). Wrapping the recieved string in a bytebuffer fixed it for now.