Fixed bug with server not sending binary messages.
This is actually a huge problem that only showed up in larger/more complex messages. It was pretty hard to pin down, but occassionally the browser WebSock client would simply crash with error code 1006 (abornmally terminated) without any further explanation. A lot of research and some reading of the RFC6455 WebSock protocol standard lead me to believe that we are not actually recieving binary data, even though the websock client is configured to expect this - and hence crashes. The WebSock library that our server uses actually defaults to setting the 4 bit opcode of the Websock protocol message frame to 129 (Fragment/Text) - not what we want.
With this commit we are now sending messages with opcode 130 (fragment/binary), and the bug appears to be fixed.