Added card folding capabilities to tables.
This is done in order to fix a bug in the previous commit, where we are basically preparing to send each player all the card info, thereby defeating the purpose of card obfuscation. Instead, we need to send players only the obfuscationTable for their own cards.

Table.hpp/cpp:
 - Added fold templates for player1 and 2 cards. These fold an accumulator through all the players cards, including crew, space, systems and hand
 - Added player1ObfuscationTableMessage methods (and player2) that use these folds to accumulate a protobuf ObfuscationTableMessage geared to only one players cards.

Table.proto:
 - Added crew and space, which we needed anyway but should not forget in folding.