Fixed deadlock when sending multiple server messages in the same block. (Note: not really fixed, just made less likely)
In PlayModule::sendStartTableMessage 2 Server messages were created from the ServerMessageFactory, which recycles messages and has a dynamically resizing arena for reusable messages.
When the arena is used up, but more messages are requested, it locks all mutexes and waits for all messages to come home (be destroyed) in order to double its arena size and reallocate.
This led to a deadlock since the starting arena size was set to 1 (one message).
In the aforementioned function 2 messages were created: the first one takes up the slot, upon the second more messages are requested and the arena starts blocking. However, the message created efore is not destroyed until the function is exited, which stalls on the blocking arena message creation request - hence deadlock.

This is currently fixed by increasing the default arena size of server messages to 32 at startup. However the problem still persists on principle. Only now we can make up to 32 messages in one block before we have to free some of them. This should be ok for most uses, but we certainly have to keep an eye on this.

GameServerImpl.cpp:
 - Increased default ServerMessage arena size to 32 at startup

PlayMode.cpp:
 - Some additional logging (used to track down this deadlock)